refactor(asic): move peripherals into bm13xx
Move the reset line trait and the board peripherals bundle
from the generic hash thread module to the BM13xx module.
Only the BM13xx thread and the boards that build one use
them. The CPU thread takes no peripherals, and no signature
in the hash thread trait mentions them; they sat in the
generic module by location, not by contract. A thread for
another chip line can generalize them again when it exists.
